move regex to helper and fix ot
