+++ lib/api/helpers.js
@@ -1,6 +1,6 @@
 import url from "url";
 import { Meteor } from "meteor/meteor";
-import { check } from "meteor/check";
+import { check, Match } from "meteor/check";
 import { Router } from "/imports/plugins/core/router/lib";
 import { Shops } from "/lib/collections";
 
@@ -367,3 +367,30 @@ export function convertWeight(from, to, weight) {
   // if we made it here, something has gone wrong
   throw new Meteor.Error("invalid-parameter", "Invalid from or to value specified");
 }
+
+/**
+ * @name luhnValid
+ * @method
+ * @memberOf Helpers
+ * @summary Checks if a number passes Luhn's test
+ * @param {String} cardNumber The card number to check
+ * @returns {Boolean} The result of the test
+ */
+function luhnValid(cardNumber) {
+  return [...cardNumber].reverse().reduce((sum, c, i) => {
+    let d = parseInt(c, 10);
+    if (i % 2 !== 0) { d *= 2; }
+    if (d > 9) { d -= 9; }
+    return sum + d;
+  }, 0) % 10 === 0;
+}
+
+// Regex to do card validations
+export const ValidCardNumber = Match.Where((x) => /^[0-9]{12,19}$/.test(x) && luhnValid(x));
+
+export const ValidExpireMonth = Match.Where((x) => /^[0-9]{1,2}$/.test(x));
+
+export const ValidExpireYear = Match.Where((x) => /^[0-9]{4}$/.test(x));
+
+export const ValidCVV = Match.Where((x) => /^[0-9]{3,4}$/.test(x));
+
